+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ New features
 ============
 
 * Add modeling.get_thetas, modeling.get_omegas and modeling.get_sigmas
+* Add configuration option for NONMEM license file path
 
 Bugfixes
 ========